taquy-vm/backup/requirements.txt
schedule==1.1.0
boto3==1.17.77
python-crontab==2.5.1
croniter==1.0.13